chore: add more tests
